+++ packages/syside-languageserver/src/node/node-file-system-provider.ts
@@ -18,11 +18,36 @@ import { NodeFileSystemProvider } from "langium/node";
 import { SysMLFileSystemProvider } from "../services/shared/workspace/file-system-provider";
 import fs from "fs";
 import { URI } from "vscode-uri";
+import { pathToURI, resolvePathURI } from "syside-base";
+import { backtrackToDirname } from "../utils";
 
 export class SysMLNodeFileSystemProvider
     extends NodeFileSystemProvider
     implements SysMLFileSystemProvider
 {
+    protected stdlib: URI | undefined;
+    private _extensionDir: URI | undefined | null;
+
+    get standardLibrary(): URI | undefined {
+        return this.stdlib;
+    }
+
+    /**
+     * Extension root directory
+     */
+    get extensionDir(): URI | undefined {
+        if (this._extensionDir === undefined) {
+            const extensionDir = backtrackToDirname(__dirname, /out|src/);
+            this._extensionDir = extensionDir ? pathToURI(extensionDir) : null;
+        }
+
+        return this._extensionDir ?? undefined;
+    }
+
+    updateStandardLibrary(value: string): void {
+        this.stdlib = value ? resolvePathURI(value) : undefined;
+    }
+
     async exists(path: URI): Promise<boolean> {
         return fs.promises.stat(path.fsPath).then(
             /* resolved */ () => true,
@@ -32,6 +57,15 @@ export class SysMLNodeFileSystemProvider
 
     existsSync(path: URI): boolean {
         return fs.existsSync(path.fsPath);
+    }
+
+    async loadScript(path: URI): Promise<object | undefined> {
+        return import(path.fsPath);
+    }
+
+    async preloadFiles(): Promise<void> {
+        // nothing to preload here
+        return;
     }
 }