+++ packages/syside-vscode/src/browser/vscode.ts
@@ -14,8 +14,24 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 
-export * from "./evaluator";
-export * from "./extension-manager";
-export * from "./model-utils";
-export * from "./service-registry";
-export * from "./workspace";
+import { ClientConfig, MaybePromise } from "syside-languageclient";
+import { BaseSysMLVSCodeClientExtender } from "../common/vscode";
+
+/**
+ * SysML language client for browser VS Code.
+ */
+export class SysMLVSCodeClientExtender extends BaseSysMLVSCodeClientExtender {
+    protected override selectStdlibPath(): MaybePromise<string | undefined> {
+        // language server uses virtual FS so there is nothing to select here
+        return this.stdlibURL;
+    }
+    protected override loadConfig(): MaybePromise<ClientConfig | undefined> {
+        return this.config;
+    }
+    protected override saveConfig(): MaybePromise<void> {
+        return;
+    }
+    protected override maybeUpdateDownloadedStdlib(): MaybePromise<void> {
+        return;
+    }
+}